Use Nuke processor for cropping borders
